requirements.txt
PyMuPDF==1.24.7
lightgbm==4.4.0
pillow==10.4.0
google-generativeai==0.7.1
rapidfuzz==3.9.3
numpy==1.26.4
opencv-python==4.10.0.84
python-dotenv==1.0.1
scikit-learn==1.5.1
pandas==2.2.2
joblib==1.4.2
nltk==3.8.1